requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.2
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.10
django-templated-mail==1.1.1
djangorestframework==3.11.1
djoser==2.0.5
flake8==3.9.2
idna==3.2
importlib-metadata==4.6.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mccabe==0.6.1
openapi-codec==1.3.2
packaging==21.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
simplejson==3.17.3
sqlparse==0.4.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
zipp==3.5.0
